修改流水线任务执行策略 从datax服务回调到arrange减少复杂度

+++ arrange/src/main/java/com/cyj/arrange/bean/vo/TCfgPipelineTaskVO.java
@@ -5,5 +5,5 @@ import lombok.Data;
 
 @Data
 public class TCfgPipelineTaskVO extends TCfgTask {
-    private Long ptid;
+    private Integer ptid;
 }
